refactor: reduce the number of times we loop through the apps in the DST
