+++ internal/app/utils.go
@@ -17,9 +17,6 @@ import (
 	"strings"
 	"time"
 
-	"gopkg.in/yaml.v2"
-
-	"github.com/BurntSushi/toml"
 	"github.com/Praqma/helmsman/internal/aws"
 	"github.com/Praqma/helmsman/internal/azure"
 	"github.com/Praqma/helmsman/internal/gcs"
@@ -36,153 +33,6 @@ func printMap(m map[string]string, indent int) {
 func printNamespacesMap(m map[string]*namespace) {
 	for key, value := range m {
 		fmt.Println(key, " : protected = ", value)
-	}
-}
-
-// fromTOML reads a toml file and decodes it to a state type.
-// It uses the BurntSuchi TOML parser which throws an error if the TOML file is not valid.
-func fromTOML(file string, s *state) (bool, string) {
-	rawTomlFile, err := ioutil.ReadFile(file)
-	if err != nil {
-		return false, err.Error()
-	}
-
-	tomlFile := string(rawTomlFile)
-	if !flags.noEnvSubst {
-		if ok, err := validateEnvVars(tomlFile, file); !ok {
-			return false, err
-		}
-		tomlFile = substituteEnv(tomlFile)
-	}
-	if !flags.noSSMSubst {
-		tomlFile = substituteSSM(tomlFile)
-	}
-	if _, err := toml.Decode(tomlFile, s); err != nil {
-		return false, err.Error()
-	}
-	resolvePaths(file, s)
-	substituteVarsInStaticFiles(s)
-
-	return true, "Parsed TOML [[ " + file + " ]] successfully and found [ " + strconv.Itoa(len(s.Apps)) + " ] apps"
-}
-
-// toTOML encodes a state type into a TOML file.
-// It uses the BurntSuchi TOML parser.
-func toTOML(file string, s *state) {
-	log.Info("Printing generated toml ... ")
-	var buff bytes.Buffer
-	var (
-		newFile *os.File
-		err     error
-	)
-
-	if err := toml.NewEncoder(&buff).Encode(s); err != nil {
-		log.Fatal(err.Error())
-		os.Exit(1)
-	}
-	newFile, err = os.Create(file)
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	bytesWritten, err := newFile.Write(buff.Bytes())
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	log.Info(fmt.Sprintf("Wrote %d bytes.\n", bytesWritten))
-	newFile.Close()
-}
-
-// fromYAML reads a yaml file and decodes it to a state type.
-// parser which throws an error if the YAML file is not valid.
-func fromYAML(file string, s *state) (bool, string) {
-	rawYamlFile, err := ioutil.ReadFile(file)
-	if err != nil {
-		return false, err.Error()
-	}
-
-	yamlFile := string(rawYamlFile)
-	if !flags.noEnvSubst {
-		if ok, err := validateEnvVars(yamlFile, file); !ok {
-			return false, err
-		}
-		yamlFile = substituteEnv(yamlFile)
-	}
-	if !flags.noSSMSubst {
-		yamlFile = substituteSSM(yamlFile)
-	}
-
-	if err = yaml.UnmarshalStrict([]byte(yamlFile), s); err != nil {
-		return false, err.Error()
-	}
-	resolvePaths(file, s)
-	substituteVarsInStaticFiles(s)
-
-	return true, "Parsed YAML [[ " + file + " ]] successfully and found [ " + strconv.Itoa(len(s.Apps)) + " ] apps"
-}
-
-// toYaml encodes a state type into a YAML file
-func toYAML(file string, s *state) {
-	log.Info("Printing generated yaml ... ")
-	var buff bytes.Buffer
-	var (
-		newFile *os.File
-		err     error
-	)
-
-	if err := yaml.NewEncoder(&buff).Encode(s); err != nil {
-		log.Fatal(err.Error())
-		os.Exit(1)
-	}
-	newFile, err = os.Create(file)
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	bytesWritten, err := newFile.Write(buff.Bytes())
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	log.Info(fmt.Sprintf("Wrote %d bytes.\n", bytesWritten))
-	newFile.Close()
-}
-
-// substituteVarsInStaticFiles loops through the values/secrets files and substitutes variables into them.
-func substituteVarsInStaticFiles(s *state) {
-	for _, v := range s.Apps {
-		if v.ValuesFile != "" {
-			v.ValuesFile = substituteVarsInYaml(v.ValuesFile)
-		}
-		if v.SecretsFile != "" {
-			v.SecretsFile = substituteVarsInYaml(v.SecretsFile)
-		}
-
-		if len(v.Hooks) != 0 {
-			for key, val := range v.Hooks {
-				if key != "deleteOnSuccess" && key != "successTimeout" && key != "successCondition" {
-					hook := val.(string)
-					if err := isValidFile(hook, []string{".yaml", ".yml"}); err != nil {
-						v.Hooks[key] = substituteVarsInYaml(hook)
-					}
-				}
-			}
-		}
-
-		if len(s.Settings.GlobalHooks) != 0 {
-			for key, val := range s.Settings.GlobalHooks {
-				if key != "deleteOnSuccess" && key != "successTimeout" && key != "successCondition" {
-					hook := val.(string)
-					if err := isValidFile(hook, []string{".yaml", ".yml"}); err != nil {
-						s.Settings.GlobalHooks[key] = substituteVarsInYaml(hook)
-					}
-				}
-			}
-		}
-
-		for i := range v.ValuesFiles {
-			v.ValuesFiles[i] = substituteVarsInYaml(v.ValuesFiles[i])
-		}
-		for i := range v.SecretsFiles {
-			v.SecretsFiles[i] = substituteVarsInYaml(v.SecretsFiles[i])
-		}
 	}
 }
 
@@ -224,74 +74,6 @@ func stringInSlice(a string, list []string) bool {
 		}
 	}
 	return false
-}
-
-// resolvePaths resolves relative paths of certs/keys/chart/value file/secret files/etc and replace them with a absolute paths
-func resolvePaths(relativeToFile string, s *state) {
-	dir := filepath.Dir(relativeToFile)
-	downloadDest, _ := filepath.Abs(createTempDir(tempFilesDir, "tmp"))
-	for _, v := range s.Apps {
-		if v.ValuesFile != "" {
-			v.ValuesFile, _ = resolveOnePath(v.ValuesFile, dir, downloadDest)
-		}
-		if v.SecretsFile != "" {
-			v.SecretsFile, _ = resolveOnePath(v.SecretsFile, dir, downloadDest)
-		}
-
-		if len(v.Hooks) != 0 {
-			for key, val := range v.Hooks {
-				if key != "deleteOnSuccess" && key != "successTimeout" && key != "successCondition" {
-					hook := val.(string)
-					if err := isValidFile(hook, []string{".yaml", ".yml", ".json"}); err != nil {
-						v.Hooks[key], _ = resolveOnePath(hook, dir, downloadDest)
-					}
-				}
-			}
-		}
-
-		for i := range v.ValuesFiles {
-			v.ValuesFiles[i], _ = resolveOnePath(v.ValuesFiles[i], dir, downloadDest)
-		}
-		for i := range v.SecretsFiles {
-			v.SecretsFiles[i], _ = resolveOnePath(v.SecretsFiles[i], dir, downloadDest)
-		}
-
-		if v.Chart != "" {
-			var repoOrDir = filepath.Dir(v.Chart)
-			_, isRepo := s.HelmRepos[repoOrDir]
-			isRepo = isRepo || stringInSlice(repoOrDir, s.PreconfiguredHelmRepos)
-			if !isRepo {
-				// if there is no repo for the chart, we assume it's intended to be a local path
-
-				// support env vars in path
-				v.Chart = os.ExpandEnv(v.Chart)
-				// respect absolute paths to charts but resolve relative paths
-				if !filepath.IsAbs(v.Chart) {
-					v.Chart, _ = filepath.Abs(filepath.Join(dir, v.Chart))
-				}
-			}
-		}
-	}
-	// resolving paths for Bearer Token path in settings
-	if s.Settings.BearerTokenPath != "" {
-		s.Settings.BearerTokenPath, _ = resolveOnePath(s.Settings.BearerTokenPath, dir, downloadDest)
-	}
-	// resolve paths for global hooks
-	if len(s.Settings.GlobalHooks) != 0 {
-		for key, val := range s.Settings.GlobalHooks {
-			if key != "deleteOnSuccess" && key != "successTimeout" && key != "successCondition" {
-				hook := val.(string)
-				if err := isValidFile(hook, []string{".yaml", ".yml", ".json"}); err != nil {
-					s.Settings.GlobalHooks[key], _ = resolveOnePath(hook, dir, downloadDest)
-				}
-			}
-		}
-	}
-	// resolving paths for k8s certificate files
-	for k := range s.Certificates {
-		s.Certificates[k], _ = resolveOnePath(s.Certificates[k], "", downloadDest)
-	}
-
 }
 
 // resolveOnePath takes the input file (URL, cloud bucket, or local file relative path),
